% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADdata.R
\docType{data}
\name{ADdata}
\alias{ADdata}
\title{AD dataset}
\format{
A data frame with X rows and Y columns:
\describe{
\item{column1}{Description of column1.}
\item{column2}{Description of column2.}
\item{...}{...}
}
}
\source{
Gene Expression Omnibus (GEO) under the accession number GSE174367
}
\usage{
ADdata
}
\description{
A dataset containing information on ...
}
\keyword{datasets}
