% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estParams.R
\name{estParams}
\alias{estParams}
\title{estimating ZINB parameters}
\usage{
estParams(object, group.1 = NULL, group.2 = NULL)
}
\arguments{
\item{object}{scaDAdataset object}

\item{group.1}{cell type 1}

\item{group.2}{cell type 2}
}
\value{
a scaDAdataset object contains parameter estimates in params slot
}
\description{
estimating ZINB parameters
}
