% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estParams2.R
\name{estParams2}
\alias{estParams2}
\title{estimating ZINB parameters}
\usage{
estParams2(object, contrl.loc, case.loc)
}
\arguments{
\item{object}{scaDAdataset object}

\item{contrl.loc}{cell type of interest}

\item{case.loc}{cell type of interest}
}
\value{
a scaDAdataset object contains parameter estimates in params slot
}
\description{
estimating ZINB parameters
}
