% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hbrain.R
\docType{data}
\name{Hbrain}
\alias{Hbrain}
\title{Human Brian dataset}
\format{
A data frame with X rows and Y columns:
\describe{
\item{column1}{Description of column1.}
\item{column2}{Description of column2.}
\item{...}{...}
}
}
\source{
10X Genomics
}
\usage{
Hbrain
}
\description{
A dataset containing information on ...
}
\keyword{datasets}
