% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanBrain.R
\docType{data}
\name{HumanBrain}
\alias{HumanBrain}
\title{Human Brian dataset}
\format{
A data frame with X rows and Y columns:
\describe{
\item{column1}{Description of column1.}
\item{column2}{Description of column2.}
\item{...}{...}
}
}
\source{
10X Genomics
}
\usage{
HumanBrain
}
\description{
A dataset containing information on ...
}
\keyword{datasets}
