% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optParams.R
\name{optParams}
\alias{optParams}
\title{Title}
\usage{
optParams(object)
}
\arguments{
\item{object}{scaDAdataset object}
}
<<<<<<< HEAD
\value{
scaDAdataset object
}
=======
>>>>>>> a03dfd55acdc9b6e8f98bf836fdd0a1457b02406
\description{
Title
}
