% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optParamsParallel.R
\name{optParamsParallel}
\alias{optParamsParallel}
\title{optimization of mu and phi parameters}
\usage{
optParamsParallel(object)
}
\arguments{
\item{object}{scaDAdataset object}
}
\value{
scaDAdataset object
}
\description{
optimization of mu and phi parameters
}
