% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{scaDAdataset-class}
\alias{scaDAdataset-class}
\title{S4 class for storing scaDA dataset}
\description{
This S4 class is designed to store the count matrix and associated column data information
for the scaDA package. currently the package can only applied to paired experiment design, single factor
}
\section{Slots}{

\describe{
\item{\code{data}}{A matrix containing the read counts from experiments, must be positive integers}

\item{\code{colData}}{A data.frame containing column metadata associated with experiment design.}
}}

