% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{scaDAdatasetFromMatrix}
\alias{scaDAdatasetFromMatrix}
\title{scaDAdataset object and constructors; currently it need to be created from matrix
later we will make it seamlessly integerated with seurat package.
Constructor for scaDAdataset}
\usage{
scaDAdatasetFromMatrix(count, colData)
}
\arguments{
\item{count}{A matrix containing the dataset.}

\item{colData}{A data.frame containing column metadata.}
}
\value{
An object of class scaDAdataset.
}
\description{
scaDAdataset object and constructors; currently it need to be created from matrix
later we will make it seamlessly integerated with seurat package.
Constructor for scaDAdataset
}
