% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkDisp.R
\name{shrinkDisp}
\alias{shrinkDisp}
\title{Title}
\usage{
shrinkDisp(object)
}
\arguments{
\item{object}{scaDAdataset object}
}
\value{
scaDAdataset object containing shrinked dispersion estimates
}
\description{
Title
}
