% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estParams.R
\name{estParams}
\alias{estParams}
\title{estimating ZINB parameters}
\usage{
estParams(object, celltype1, celltype2)
}
\arguments{
\item{object}{scaDAdataset object}

\item{celltype1}{control cell types}

\item{celltype2}{case cell types}
}
\value{
a scaDAdataset object contains parameter estimates in params slot
}
\description{
estimating ZINB parameters
}
