% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optParams.R
\name{optParams}
\alias{optParams}
\title{optimization of mu and phi parameters}
\usage{
optParams(object)
}
\arguments{
\item{object}{scaDAdataset object}
}
\value{
scaDAdataset object
}
\description{
optimization of mu and phi parameters
}
